third_party/WebKit/LayoutTests/bluetooth/requestDevice/acceptAllDevices/device-with-name.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script>
promise_test(() => {
  let device_name = 'LE Device';
  return setUpPreconnectedDevice({name: device_name})
    .then(() => requestDeviceWithKeyDown({acceptAllDevices: true}))
    .then(device => {
      assert_equals(device.name, device_name);
    });
}, 'A device with name and no UUIDs nearby. Should be found if ' +
   'acceptAllDevices is true.');
</script>
